chore: split commands file into protocol-specific files
