+++ packages/types/lib/commands/index.ts
@@ -1,0 +1,6 @@
+export * from './appium';
+export * from './basedriver';
+export * from './jsonwp';
+export * from './mjsonwp';
+export * from './other-protocols';
+export * from './webdriver';
